+++ src/plugins/network/SftpFileSystem.h
@@ -49,7 +49,7 @@ public:
     Result<void> rename(const VfsUri& from, const VfsUri& to) override;
 
     Result<std::unique_ptr<QIODevice>> openRead(const VfsUri& target, qint64 expectedSize = -1) override;
-    Result<std::unique_ptr<QIODevice>> openWrite(const VfsUri& target) override;
+    Result<std::unique_ptr<QIODevice>> openWrite(const VfsUri& target, qint64 expectedSize = -1) override;
 
     Result<AccessInfo> access(const VfsUri& target) override;